update payment method UI; push to prod
