+++ src/components/Suppliers/SupplierDashboard/SupplierDashboard.js
@@ -1,62 +1,60 @@
 import React, { useState, useEffect } from "react";
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import axios from "axios";
-
-/**
- * still not work with cookie
- */
-// Function to get the value of a cookie by name
-function getCookie(name) {
-  const cookies = document.cookie.split("; ");
-  for (const cookie of cookies) {
-    const [cookieName, cookieValue] = cookie.split("=");
-    if (cookieName === name) {
-      return decodeURIComponent(cookieValue);
-    }
-  }
-  return null;
-}
+import "./SupplierDashboard.css";
 
 const SupplierDashboard = () => {
   const [supplierData, setSupplierData] = useState({});
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
   const { supplierId } = useParams();
+  const [paypalToken, setPaypalToken] = useState(null);
+  const navigation = useNavigate(); // Initialize useHistory hook
 
-  // Get the value of the "authToken" cookie
-  const authToken = getCookie("authToken");
-  console.log("authToken: " + authToken)
+  const pageurl = "/assets/img/umiuni/logo/logo300x83.png";
 
   useEffect(() => {
+    setLoading(true);
+    const token = localStorage.getItem("token");
+  
     const fetchSupplierData = async () => {
+      const queryParams = new URLSearchParams(window.location.search);
       try {
         const response = await axios.get(
+          // `${process.env.REACT_APP_API_URL}/api/v1/suppliers/${supplierId}`,
           `${process.env.REACT_APP_API_URL}/api/v1/suppliers/auth/info`,
           {
             headers: {
-              Authorization: `Bearer ${authToken}`,
+              Authorization:
+                `Bearer ` +
+                (localStorage.getItem("token") || queryParams.get("token")), // `Bearer ${token}`, //`Bearer ` + queryParams.get("token"), // token Attention: here will get the network error if use token
             },
-            withCredentials: true, // Include credentials (cookies)
           }
         );
-        console.log("response.data: " + response.data)
         setSupplierData(response.data);
       } catch (err) {
-        setError(err.message);
+        setError(err.message + ",  Please Login First!");
       } finally {
         setLoading(false);
       }
     };
 
-    if (authToken) {
-      // Fetch supplier data only if the "authToken" cookie exists
-      fetchSupplierData();
-    } else {
-      // Handle the case where the "authToken" cookie is not found
-      setLoading(false);
-    }
-  }, [supplierId, authToken]);
+    fetchSupplierData();
+  }, [supplierId]);
 
+  // Retrieve and store the PayPal token when the component mounts
+  useEffect(() => {
+    const queryParams = new URLSearchParams(window.location.search);
+    const paypalToken = queryParams.get("token");
+    if (paypalToken) {
+      setPaypalToken(paypalToken);
+      localStorage.setItem("token", paypalToken);
+    }
+  }, []);
+
+  const redirectToIMS = () => {
+    navigation('/supplier-ims');
+  };
 
   if (loading) {
     return <div>Loading your dashboard...</div>;
@@ -67,13 +65,30 @@ const SupplierDashboard = () => {
   }
 
   return (
-    <div>
-      <h1>Supplier Dashboard</h1>
-      {supplierData && (
-        <div>
-          <p>Welcome, {supplierData.supplierName}!</p>
+    <div className="supplier-dashboard">
+      <header className="dashboard-header">
+        <h1>Supplier Dashboard</h1>
+        {supplierData && (
+          <div className="welcome-message">
+            <p>Welcome, {supplierData.supplierName}!</p>
+          </div>
+        )}
+      </header>
+
+      <div className="balance-info">
+        <span>账户余额(余额) | Account Balance (Balance)</span>
+        <h1>¥ 1000.00</h1>
+        {/* // Add any buttons or links that are necessary */}
+      </div>
+      <div className="action-panel">
+        {/* // Add any specific actions or display elements here */}
+      </div>
+      <div className="temu-area">
+        <div className="temu-logo">
+          <img src={pageurl} alt="Logo" />
         </div>
-      )}
+        <button className="temu-button" onClick={redirectToIMS}>进入 | Enter Main Page</button>
+      </div>
     </div>
   );
 };
